+++ CHANGELOG.md
@@ -11,6 +11,10 @@ ______________________________________________________________________
 
 ## [Unreleased]
 
+### Fixed
+
+- Database factory functions (`db_SQLite(..., new_db=True)`, etc.) no longer crash with `AttributeError` when called before `run()` in library use.
+
 ______________________________________________________________________
 
 ## [2.22.0] - 2026-07-03